+++ sample-tracking-status-overview-app/src/main/groovy/life/qbic/portal/sampletracking/components/projectoverview/ProjectOverviewView.groovy
@@ -8,6 +8,7 @@ import com.vaadin.server.StreamResource
 import com.vaadin.shared.ui.ContentMode
 import com.vaadin.shared.ui.grid.HeightMode
 import com.vaadin.ui.*
+import com.vaadin.ui.Grid.Column;
 import com.vaadin.ui.themes.ValoTheme
 import groovy.util.logging.Log4j2
 import life.qbic.portal.sampletracking.Constants
@@ -198,9 +199,7 @@ class ProjectOverviewView extends VerticalLayout{
         return state.getCssClass()
     }
 
-
     private void fillProjectsGrid() {
-
         projectGrid.addColumn({ it.code })
                 .setCaption("Project Code").setId("ProjectCode").setMaximumWidth(
                 MAX_CODE_COLUMN_WIDTH)
@@ -233,6 +232,11 @@ class ProjectOverviewView extends VerticalLayout{
         projectGrid.getColumn("SampleDataAvailable")
                 .setMaximumWidth(MAX_STATUS_COLUMN_WIDTH).setExpandRatio(1)
         projectGrid.setHeightMode(HeightMode.ROW)
+        
+        // remove manual sorting - any sorting in the code should probably done before disabling it
+        for (Column col : projectGrid.getColumns()) {
+          col.setSortable(false);
+        }
     }
 
     private void setupDataProvider() {