+++ sample-tracking-status-overview-app/src/main/groovy/life/qbic/portal/sampletracking/components/projectoverview/ProjectSummary.groovy
@@ -1,7 +1,7 @@
 package life.qbic.portal.sampletracking.components.projectoverview
 
 import groovy.transform.EqualsAndHashCode
-import life.qbic.datamodel.dtos.projectmanagement.Project
+import java.time.Instant
 import life.qbic.portal.sampletracking.components.projectoverview.statusdisplay.SampleCount
 import life.qbic.business.project.Project
 
@@ -22,6 +22,7 @@ class ProjectSummary {
     SampleCount samplesQc
     SampleCount samplesLibraryPrepFinished
     SampleCount sampleDataAvailable
+    Instant lastChanged
 
     ProjectSummary(String code, String title) {
         this.code = code
@@ -31,12 +32,15 @@ class ProjectSummary {
         this.samplesLibraryPrepFinished = new SampleCount(0,0,0)
         this.sampleDataAvailable = new SampleCount(0,0,0)
         this.totalSampleCount = 0
+        this.lastChanged = Instant.MIN
     }
 
     static ProjectSummary of(Project project) {
         String code = project.code
         String title = project.title
-        return new ProjectSummary(code, title)
+        ProjectSummary summary = new ProjectSummary(code, title)
+        summary.lastChanged = project.lastChanged
+        return summary
     }
 
     String getCode() {