+++ CHANGELOG.rst
@@ -4,8 +4,21 @@ Changelog
 
 This project adheres to `Semantic Versioning <https://semver.org/>`_.
 
+0.7.0-SNAPSHOT (2021-10-05)
 
-0.6.0-SNAPSHOT (2021-09-27)
+**Added**
+
+* Removes manual sorting option from project grid
+* Projects are now sorted by the last change as found in the sample tracking database  (`#114 <https://github.com/qbicsoftware/sample-tracking-status-overview/pull/114>`_)
+
+**Fixed**
+
+**Dependencies**
+
+**Deprecated**
+
+
+0.6.0-SNAPSHOT (2021-09-30)
 ---------------------------
 
 **Added**